requirements.txt
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
python-dotenv==1.0.1
requests==2.32.3